Refactor Cards View
