+++ src/webapp/src/components/Navigation/index.js
@@ -31,7 +31,7 @@ export default function Navigation() {
         width: '100%',
         position: 'fixed',
         bottom: '0px',
-        paddingBottom: '10px',
+        height: '65px',
       }}
     >
       <BottomNavigationAction
